+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+## Version 1.0.0
+> - Added async callback methods and releasing polished version.
+> - When you subscribe you get the current state from a store.
+
 ## Version 0.1.27
 > - Substore identifiers can be any string, removing restrictions.
 ## Version 0.1.26